Refactor triton heuristics backend registry
Motivation:
1. Current device backends share one codepath, forcing ad-hoc `if device_type == ...` handling.
2. CUDA-only config changes have been applied to XPU, causing XPU performance regressions.
3. `triton_heuristics.py` has grown too large and hard to maintain.

Design:
Split device-specific heuristic implementations into per-backend modules and route through a registry, with shared logic consolidated in a common helper module. This keeps backend differences explicit and reduces cross-device coupling.

Test Plan:
Covered by existing Inductor test cases; performance validation required.

[ghstack-poisoned]
