+++ torch/_inductor/runtime/triton_heuristics_impl/__init__.py
@@ -1,0 +1,1 @@
+from . import registry  # noqa: F401
